Fix resize issue with GraphTree where content would go down
In dojo, according to the documentation, the ContentPane is expected to be the leaf element in the layout.

Having nested ContentPane seems to trigger a weird side effect where dragging the splitter would also resize (horizontally) the paramsPane and the graphPane.

The fix is simply to use a LayoutContainer instead of ContnetPane for the content.