Fix classpath for "tests" task of ant
Sets "fork" to true to force the tests to be runned in a different JVM with the specified classpath.

Otherwise it will use the same classpath as ant and share its classpath which has its own version of junit.

In my case I have a version of ant which cause a NoSuchMethodException when trying to execute the task.

Searching over the web it seems to be a pretty common issue with ant.