avoid little flickering of tree when loading.
Because we hide the tree's root, dojo displays a little clock to indicate that we are currently loading the content but we have already a "loading" overlay on top of all the tree and because we already have content it makes everything flickers.

There is no way in JS to avoid the display of this image so let's force the display: none and no image to avoid downloading it for nothing.