make tree loading feel faster.
The standBy loading spinner has a fadeIn/fadeOut effect duration of 500ms. This is awfully long considering the time the requests take.

Let's remove this duration because we want an UI that feels responsive and that is practical. We don't really care about CSS effects that slows everything down.

Also recycle the standBy widget, not sure why it was destroyed/recreated, maybe I overlooked something.