Fix time period filter where date would change when focused (#58)
When focusing a date or time input, we automatically set the  autoperiod to '0' (meaning manual) but the comparison in 'onChange' callback would always fails and always update the input to the latest value.

The comparison fails because 'value' is a string and '!==' means a strict equality of both value and type.